Refactor router transitions and add animations to App and Login components
The code changes refactor the router transitions by adding animations to
the App and Login components. The App component now includes a zoom-in
animation, while the Login component includes a zoom-out animation. This
enhances the user experience and provides a smoother visual transition
between pages. The addition of the animate.css file facilitates the
implementation of these animations. This change improves the overall
look and feel of the application and enhances its usability.

Issue: N/A